Refactor controllers to use DTOs and update build process to use Gradle

+++ src/main/java/com/example/fedegan/dto/AnimalDTO.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AnimalDTO {
-    private int animal_id;
-    private int finca_id;
+    private Long animal_id;
+    private Long finca_id;
     private String tipo_animal;
     private String raza;
     private String sexo;
